Updated Deal script
Changed how CardDeal phase ends. Before, it would keep checking if either player, depending on turn, are holding a total of 5 cards before moving to the next phase.
Now, an event added in BattleManager will be called out once cards are dealt, allowing for the CardDeal to end and move to the next phase without any checks (except for bool)